+++ modules/kubernetes-cluster/src/main.tf
@@ -52,6 +52,14 @@ resource "azurerm_kubernetes_cluster" "main" {
     }
   }
 
+  dynamic "windows_profile" {
+    for_each = var.windows_profile != null ? [var.windows_profile] : []
+    content {
+      admin_password = windows_profile.value.admin_password
+      admin_username = windows_profile.value.admin_username
+    }
+  }
+
   key_vault_secrets_provider {
     secret_rotation_enabled = true
   }