+++ modules/vpn-server-configuration/src/terraform.tf
@@ -2,6 +2,10 @@ terraform {
   required_version = "~> 1.5"
 
   required_providers {
+    azuread = {
+      source  = "hashicorp/azuread"
+      version = "~> 3.1"
+    }
     azurerm = {
       source  = "hashicorp/azurerm"
       version = "~> 4.23"