+++ scripts/helpers/lock_terraform_providers.py
@@ -1,4 +1,4 @@
-from python_terraform import Terraform
+from python_terraform import IsFlagged, Terraform
 
 
 def lock_terraform_providers(working_dir: str, **kwargs: dict):
@@ -6,7 +6,10 @@ def lock_terraform_providers(working_dir: str, **kwargs: dict):
 
     terraform = Terraform(working_dir=working_dir)
     return_code, _, _ = terraform.cmd(
-        "providers lock", platform=["darwin_amd64", "linux_amd64"], capture_output=False
+        "providers lock",
+        platform=["darwin_arm64", "linux_amd64"],
+        enable_plugin_cache=IsFlagged,
+        capture_output=False,
     )
     if return_code != 0:
         exit(return_code)