+++ modules/data-factory/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+## 2.0
+* Updated the AzureRM Terraform Provider from 3.x to 4.x.
+* Refactored `identity` and `vsts_configuration` variables.