+++ modules/firewall-policy/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+## 2.0
+* Updated the AzureRM Terraform Provider from 3.x to 4.x.
+* Replaced `dns_servers` variable with `dns`.