+++ modules/log-analytics-workspace/CHANGELOG.md
@@ -1,1 +1,4 @@
 # Changelog
+
+## 2.0
+* Updated the AzureRM Terraform Provider from 3.x to 4.x.