+++ modules/mysql-flexible-server/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
-## 1.1
+## 2.0
+* Updated the AzureRM Terraform Provider from 3.x to 4.x.
+* Updated the AzAPI Terraform Provider from 1.x to 2.x.
 
+## 1.1
 * Added option to control public network access.