+++ modules/postgresql-flexible-server/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 2.0
+* Updated the AzureRM Terraform Provider from 3.x to 4.x.
+
 ## 1.1
 * Changed `private_dns_zone_id` and `subnet_id` variables to be optional to support Private Endpoint usage.