+++ modules/private-dns-resolver/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 3.0
+* Updated the AzureRM Terraform Provider from 3.x to 4.x.
+
 ## 2.0
 * Changed naming convention for inbound and outbound endpoints.