+++ modules/virtual-network/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 3.0
+* Updated the AzureRM Terraform Provider from 3.x to 4.x.
+
 ## 2.0
 * Removed `dns_servers` variable to enable the use of the [azurerm_virtual_network_dns_servers](https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/virtual_network_dns_servers) resource instead.